+++ src/Profiler/Mutation/Http.php
@@ -9,6 +9,10 @@ use Innmind\Filesystem\{
     File,
     File\Content,
 };
+use Innmind\Immutable\{
+    Attempt,
+    SideEffect,
+};
 
 final class Http
 {
@@ -26,23 +30,31 @@ final class Http
         return new self($storage, $profile);
     }
 
-    public function received(Content $request): void
+    /**
+     * @return Attempt<SideEffect>
+     */
+    #[\NoDiscard]
+    public function received(Content $request): Attempt
     {
-        $_ = $this->storage->add($this->profile->add(
+        return $this->storage->add($this->profile->add(
             Directory::named('http')->add(File::named(
                 'request.txt',
                 $request,
             )),
-        ))->unwrap();
+        ));
     }
 
-    public function respondedWith(Content $response): void
+    /**
+     * @return Attempt<SideEffect>
+     */
+    #[\NoDiscard]
+    public function respondedWith(Content $response): Attempt
     {
-        $_ = $this->storage->add($this->profile->add(
+        return $this->storage->add($this->profile->add(
             Directory::named('http')->add(File::named(
                 'response.txt',
                 $response,
             )),
-        ))->unwrap();
+        ));
     }
 }